+++ payment-app/application/src/test/scala/jp/co/tis/lerna/payment/application/ecpayment/issuing/actor/PaymentActorSpec.scala
@@ -1,6 +1,6 @@
 package jp.co.tis.lerna.payment.application.ecpayment.issuing.actor
 
-import akka.actor.{ ActorRef, ActorSystem, Props, Status }
+import akka.actor.{ ActorRef, ActorSystem, Props }
 import akka.testkit.{ ImplicitSender, TestKit }
 import com.typesafe.config.{ Config, ConfigFactory }
 import jp.co.tis.lerna.payment.adapter.ecpayment.model.{ OrderId, WalletShopId }
@@ -25,10 +25,12 @@ import lerna.util.tenant.Tenant
 import lerna.util.time.{ FixedLocalDateTimeFactory, LocalDateTimeFactory }
 import lerna.util.trace.TraceId
 import org.scalatest.concurrent.ScalaFutures
-import org.scalatest.{ Inside, MustMatchers, WordSpecLike }
+import org.scalatest.Inside
 import wvlet.airframe.Design
 
 import scala.concurrent.Future
+import org.scalatest.matchers.must.Matchers
+import org.scalatest.wordspec.AnyWordSpecLike
 
 // Lint回避のため
 @SuppressWarnings(
@@ -43,8 +45,8 @@ import scala.concurrent.Future
 class PaymentActorSpec
     extends TestKit(ActorSystem("PaymentActorSpec"))
     with ImplicitSender
-    with WordSpecLike
-    with MustMatchers
+    with AnyWordSpecLike
+    with Matchers
     with DISessionSupport
     with JDBCSupport
     with Inside
@@ -145,11 +147,7 @@ class PaymentActorSpec
             AtLeastOnceDelivery.tellTo(actor, payRequest)
 
             val expect = NotFound("決済情報")
-            inside(expectMsgType[Status.Failure]) {
-              case Status.Failure(ex: BusinessException) =>
-                ex.message.messageContent must be(expect.messageContent)
-                ex.message.messageId must be(expect.messageId)
-            }
+            expectMsg(SettlementFailureResponse(expect))
           }
         }
 
@@ -182,11 +180,7 @@ class PaymentActorSpec
 
             AtLeastOnceDelivery.tellTo(actor, payRequest)
             val expect = NotFound("決済情報")
-            inside(expectMsgType[Status.Failure]) {
-              case Status.Failure(ex: BusinessException) =>
-                ex.message.messageContent must be(expect.messageContent)
-                ex.message.messageId must be(expect.messageId)
-            }
+            expectMsg(SettlementFailureResponse(expect))
           }
         }
 
@@ -264,20 +258,13 @@ class PaymentActorSpec
 
               AtLeastOnceDelivery.tellTo(actor, payRequest)
               val expect = IssuingServiceServerError("承認売上送信", "TW001")
-              inside(expectMsgType[Status.Failure]) {
-                case Status.Failure(ex: BusinessException) =>
-                  ex.message.messageContent must be(expect.messageContent)
-                  ex.message.messageId must be(expect.messageId)
-              }
+              expectMsg(SettlementFailureResponse(expect))
 
               val cancelRequest = Cancel(ClientId(777), CustomerId("789"), WalletShopId("123"), OrderId("456"))
               AtLeastOnceDelivery.tellTo(actor, cancelRequest)
               val expectedClientErrorMessage = ValidationFailure("walletShopId または orderId が不正です")
 
-              inside(expectMsgType[Status.Failure]) {
-                case Status.Failure(ex: BusinessException) =>
-                  ex.message must be(expectedClientErrorMessage)
-              }
+              expectMsg(SettlementFailureResponse(expectedClientErrorMessage))
             }
           }
         }
@@ -324,11 +311,7 @@ class PaymentActorSpec
           AtLeastOnceDelivery.tellTo(actor, payRequest)
 
           val expect = TimeOut("Some error")
-          inside(expectMsgType[Status.Failure]) {
-            case Status.Failure(ex: BusinessException) =>
-              ex.message.messageContent must be(expect.messageContent)
-              ex.message.messageId must be(expect.messageId)
-          }
+          expectMsg(SettlementFailureResponse(expect))
         }
 
         "決済失敗: 未知の異常" in withJDBC { db =>
@@ -370,22 +353,14 @@ class PaymentActorSpec
                 dateTime,
                 transactionIdFactory,
                 paymentIdFactory,
-              ) {
-                override def sendToSelf(message: InnerCommand): Unit = {
-                  AtLeastOnceDelivery.tellTo(self, message)
-                }
-              },
+              ),
             ),
             name = MultiTenantShardingSupportTestHelper.generateActorName(),
           )
 
           AtLeastOnceDelivery.tellTo(actor, payRequest)
           val expect = UnpredictableError()
-          inside(expectMsgType[Status.Failure]) {
-            case Status.Failure(ex: BusinessException) =>
-              ex.message.messageContent must be(expect.messageContent)
-              ex.message.messageId must be(expect.messageId)
-          }
+          expectMsg(SettlementFailureResponse(expect))
         }
       }
     }
@@ -473,10 +448,7 @@ class PaymentActorSpec
 
         AtLeastOnceDelivery.tellTo(actor, payRequest)
         val expectedClientErrorMessage = ValidationFailure("walletShopId または orderId が不正です")
-        inside(expectMsgType[Status.Failure]) {
-          case Status.Failure(ex: BusinessException) =>
-            ex.message must be(expectedClientErrorMessage)
-        }
+        expectMsg(SettlementFailureResponse(expectedClientErrorMessage))
       }
 
       "決済取消送信、DBに合うデータがない、永続化しない" in withJDBC { db =>
@@ -517,11 +489,7 @@ class PaymentActorSpec
         AtLeastOnceDelivery.tellTo(actor, cancelRequest)
 
         val expect = NotFound("決済情報")
-        inside(expectMsgType[Status.Failure]) {
-          case Status.Failure(ex: BusinessException) =>
-            ex.message.messageContent must be(expect.messageContent)
-            ex.message.messageId must be(expect.messageId)
-        }
+        expectMsg(SettlementFailureResponse(expect))
       }
 
       "決済取消送信、GatewayレスポンスがNormal かつ エラーコードが「既に取消済み」 → 取消失敗(IssuingServiceAlreadyCanceled)" in withJDBC { db =>
@@ -569,11 +537,7 @@ class PaymentActorSpec
         AtLeastOnceDelivery.tellTo(actor, cancelRequest)
         // CODE-011
         val expect = IssuingServiceAlreadyCanceled()
-        inside(expectMsgType[Status.Failure]) {
-          case Status.Failure(ex: BusinessException) =>
-            ex.message.messageContent must be(expect.messageContent)
-            ex.message.messageId must be(expect.messageId)
-        }
+        expectMsg(SettlementFailureResponse(expect))
       }
 
       "決済取消送信、GatewayレスポンスがNormal かつ エラーコードが正常、「既に取消済み」以外 → 取消失敗(IssuingServiceServerError" in withJDBC { db =>
@@ -619,11 +583,7 @@ class PaymentActorSpec
         val cancelRequest = Cancel(ClientId(777), CustomerId("789"), WalletShopId("123"), OrderId("456"))
         AtLeastOnceDelivery.tellTo(actor, cancelRequest)
         val expect = IssuingServiceServerError("承認取消送信", "TW900")
-        inside(expectMsgType[Status.Failure]) {
-          case Status.Failure(ex: BusinessException) =>
-            ex.message.messageContent must be(expect.messageContent)
-            ex.message.messageId must be(expect.messageId)
-        }
+        expectMsg(SettlementFailureResponse(expect))
       }
 
       "決済取消送信、GatewayレスポンスがFailure → エラー" in withJDBC { db =>
@@ -670,11 +630,7 @@ class PaymentActorSpec
         AtLeastOnceDelivery.tellTo(actor, cancelRequest)
 
         val expect = TimeOut("123")
-        inside(expectMsgType[Status.Failure]) {
-          case Status.Failure(ex: BusinessException) =>
-            ex.message.messageContent must be(expect.messageContent)
-            ex.message.messageId must be(expect.messageId)
-        }
+        expectMsg(SettlementFailureResponse(expect))
       }
 
       "取消成功、再度送信 → 取消成功" in withJDBC { db =>
@@ -761,16 +717,8 @@ class PaymentActorSpec
         AtLeastOnceDelivery.tellTo(actor, cancelRequest)
 
         val expectedMessage = gatewayException.message
-        inside(expectMsgType[Status.Failure]) {
-          case Status.Failure(ex: BusinessException) =>
-            ex.message.messageContent must be(expectedMessage.messageContent)
-            ex.message.messageId must be(expectedMessage.messageId)
-        }
-        inside(expectMsgType[Status.Failure]) {
-          case Status.Failure(ex: BusinessException) =>
-            ex.message.messageContent must be(expectedMessage.messageContent)
-            ex.message.messageId must be(expectedMessage.messageId)
-        }
+        expectMsg(SettlementFailureResponse(expectedMessage))
+        expectMsg(SettlementFailureResponse(expectedMessage))
       }
     }
 
@@ -822,16 +770,8 @@ class PaymentActorSpec
           AtLeastOnceDelivery.tellTo(actor, payRequest)
           AtLeastOnceDelivery.tellTo(actor, payRequest)
 
-          inside(expectMsgType[Status.Failure]) {
-            case Status.Failure(ex: BusinessException) =>
-              ex.message.messageContent must be(expect.messageContent)
-              ex.message.messageId must be(expect.messageId)
-          }
-          inside(expectMsgType[Status.Failure]) {
-            case Status.Failure(ex: BusinessException) =>
-              ex.message.messageContent must be(expect.messageContent)
-              ex.message.messageId must be(expect.messageId)
-          }
+          expectMsg(SettlementFailureResponse(expect))
+          expectMsg(SettlementFailureResponse(expect))
         }
       }
     }
@@ -848,11 +788,7 @@ class PaymentActorSpec
           dateTime,
           transactionIdFactory,
           paymentIdFactory,
-        ) {
-          override def sendToSelf(message: InnerCommand): Unit = {
-            AtLeastOnceDelivery.tellTo(self, message)
-          }
-        },
+        ),
       ),
       name = MultiTenantShardingSupportTestHelper.generateActorName(),
     )