+++ CHANGELOG.md
@@ -13,8 +13,18 @@ TODO: sample の version 体系について検討
 - [README](README.md) に Management APIs の使用方法を記載しました
 
 ### CHANGED
+- `lerna-app-library-2.0.0-6bad8983-SNAPSHOT` に更新しました
+    - `lerna-management` の更新に伴い、
+      次の2つの HTTP APIs は Long 値から Double 値を返すように変更します。
+        - `/metrics/rmu/sales_detail/ec_house_money/number_of_singleton`
+        - `/metrics/system-metrics/jvm-memory/heap/max`
 - `Scala 2.12.13` に更新しました
 - `sbt-wartremover 2.4.13` に更新しました
+- Akka typed 対応のため、 `PaymentActor` から `self` にメッセージを送る際の処理を変更しました
+    - graceful shutdown 時のレイテンシが増加する可能性があります
+- Akka typed 対応のため、 `PaymentActor` からのレスポンスメッセージを変更しました
+    -  `Status.Failure(exception)` -> 専用クラス化
+    - ※ Response, Event の互換性が崩れる
 
 ## Version 1.1.0
 - `Changed` Read Model Updater を分散実行しスループットを向上